requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
distro==1.9.0
Django==4.2.20
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
inflection==0.5.1
jiter==0.9.0
Markdown==3.7
openai==1.75.0
packaging==24.2
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.9.0
pytz==2025.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
uritemplate==4.1.1
zipp==3.21.0
